requirements.txt
fastapi==0.115.8
bcrypt==4.0.1
SQLAlchemy==2.0.38
kafka-python==2.0.3
psycopg2-binary==2.9.10
pydantic==2.10.6
PyJWT==2.10.1
python-dotenv==1.0.1
requests==2.32.3
uvicorn==0.34.0
